revert: Move `namespace` back
Not getting any typing benefits in `common`, so this is less complicated
